The following changes should make it now possible to properly read the addresses in the to/from/cc/bcc (etc) fields in an Outlook Express or Windows mail message.
*Added oleTypes.py which contains some useful COM interfaces for low-level OLE support (IOleObject etc).
*speech.speakSymbol now reverts to speakText if the given string is more than one character long (i.e. its not really a symbol)
*Added support to the Edit NVDAObject for embedded OLE objects. the text property in the EditTextInfo class will detect embedded objects and expand them to a useful text representation of the object.
The Edit NVDA object now has a property called editAPIHasITextDocument, which if true tells the EditTextInfo class its aloud to use ITextDocument to deal with embedded objects. Collecting embedded objects may be a little bit of a performance loss so this should only be set to True when its necessary to read embedded objects.
To get these embedded objects, NVDA must get the text using ITextDocument, then search the text for any embedded object characters, request the embedded object at that point, and then through various COM interfaces (with both comtypes and pywin32) ask for the text that would be copied to the clipboard if this object was in fact to be copied.
*Added an editValueUnit property to the edit NVDAObject. This property contains the text unit that should be used when retreaing text for the NVDAObject's value property. As in, the text you will hear when focus is set to this object. By default its set to line, but for some small multi-line fields it may be better set to story (the whole text).
*The appModule for Outlook Express / Windows Mail, if it has detected a mail envelope field (such as to/cc/bcc/from), will set editAPIHasITextDocument to True, and editValueUnit to text.UNIT_STORY, on the NVDAObject for that field.