Reimplemented support for Windows dos consoles. *Added winConsoleHandler module, which manages connection and disconnection for consoles to NVDA's process, runs a monitor thread which handles announcement of new text, and provides a TextInfo class with direct access to the current console. The code is generally the same as what was in the winConsole NVDAObject, though kernel32.GetStdHandle is no longer used, rather the CONSOLE_REAL_OUTPUT_HANDLE constant is used instead. This specific change seems to fix a bug where NVDA could not read particular consoles such as starting Cygwin ssh from the run dialog, or the Mozilla build system if alt tabbing back to it. *WinConsole Window NVDAObject: remove all original code, and set this NVDAObject to use the WinConsole TextInfo class as its TextInfo if its window matches the current console window winConsoleHandler is managing.
The new implementation (like the old one) announces new text, and allows you to navigate the content with the review cursor, and tracks the caret. Rather than speaking the entire content on gainFocus, only the line at the caret is announced, just like edit fields. Protection for control+c has not been reimplemented yet, so NVDA will terminate if it is pressed in a console.
As winConsoleHandler is separate from the winConsole NVDAObject, this provides a better possibility for reading console content under the mouse, and also allows navigating to the console with object navigation and being able to review the content. Previously you could only review the content if the navigator object was left where it was set by the focus.