Initial commit to implement a possibly better paradyme for  using both object navigation and flat review together.
When moving with object navigation, you can now press nvda+numpad7 to locate the closest object which implements a flat review, and the review cursor will try to be positioned at the original navigator object's position.

When reviewing the flat review, you can at any time press nvda+numpad1 to move the navigator object to the deepest object at the position of the review cursor.

Note that previously there was some magic that automatically moved the navigator object to the deepest object at the review cursor if the review cursor was moved after the navigator object. This no longer occures. The user now must choose to do this with nvda+numpad1.

Note that previously if the navigator object was on an object inside a treeInterceptor (virtualBuffer or document) moving the review cursor would cause you to review the entire treeInterceptor. This no longer occures, only the current object will be reviewed. The User now must choose to review the entire treeInterceptor by pressing nvda+numpad7.
   
Specific changes:
*base NVDAObject: add a flatReviewPosition property. This property fetches a TextInfo, positioned at this object, within the closest flat review. For now, this is either the object's tree interceptor (virtualBuffer, Lotus Symphony document etc), or the closest DisplayModelTextInfo.
For treeInterceptors, the TextInfo is instanciated given this object as the position. for DisplayModelTextInfo, This object's location is fetched, and then its center point is used as the point position to instanciate the TextInfo. We may want to think about allowing Offsets-based textInfos to be instanciated with an NVDAObject, but right now this will do as a prototype. 
Also because of the way _getOffsetFromPoint is implemented on DisplayModelTextInfo, not all points relate to an offset. So therefore if this happens, we just instanciate  with position_first.  
*api.getReviewPosition: If the position must be calculated from the navigator object, no longer fetch it from the navigator object's treeInterceptor if it has one.
*api.getNavigatorObject: if the object must be calculated from the review position, no longer use NVDAObjectAtStart, instead just use the TextInfo's obj if its an NVDAObject, or or the TextInfo's obj's rootNVDAObject, if its a treeInterceptor.
*Default appModule: add script navigatorObject_moveToFlatReviewAtObjectPosition (nvda+numpad7) which locates the closest flat review to the navigator object, and sets the reveiw position to the text representation of the navigator object, within the flat review.
*Default appModule: add script navigatorObject_moveToObjectAtFlatReviewPosition (nvda+numpad1) which mosets the navigator object to the deepest object represented at the current review position.