scons: Make sure the output from eSpeak dictionary compilation appears in the right place.
eSpeak dictionary compilation outputs to stderr by default, but it doesn't flush stderr once complete. This meant that some of the output sometimes appeared at the end of the scons output instead of where it actually occurred. This made build output very confusing.
Unfortunately, because we statically link the CRT in our eSpeak dll, there's no way for us to flush its stderr buffer ourselves, nor can we pass another stream because streams are handled by the CRT.
Therefore, we load and free the eSpeak dll for every dictionary compiled. Freeing the dll terminates the CRT, which flushes stderr.