Replace the current config file error dialog with a simple message box.
Users shouldn't be editing the config file in normal usage.
If the file has become corrupt, it's probably beyond fixing, so don't confuse the user by suggesting that they can fix it.
We don't validate the whole config in one shot any more, so this won't even come up if a user manually puts an incorrect value in the file.