Fix an issue where nothing was spoken if a single key (e.g. an arrow key) was pressed during say all and say all used a different speech synth.
This occurred because the say all trigger was being exited after the text in response to the key was spoken instead of before. The reason was that a reference to the say all generator was being held longer than it should in queueHandler.pumpall, so Python couldn't destroy the generator until after the queue had been pumped.
Now, pumpAll only holds the reference as long as it needs to.