ProfilesDialog: When creating a new profile, allow the user to select from a list of triggers for convenience.
The user can select Manual activation if they want to activate the profile manually rather than having any triggers.
If no name is specified, it will be filled in based on the selected trigger.
If the selected trigger needs to be manually activated in order for the user to edit it, the user will be prompted as to whether they want to do this immediately.
If appropriate, the Profiles dialog will be closed to facilitate immediate editing.