Fix more cases where a profile trigger isn't exited when an application closes.
This occurred because there were sometimes two app modules for the same process, so the profile was being triggered twice. The first instance died before the process exited, so one of the triggers wasn't exited.
The root cause was that app modules can be fetched in UIA event threads, but there was a race condition in getAppModuleFromProcessID where two app modules could be created if two threads were fetching for the same process ID at around the same time.
Fix this using a lock.
Re #3732.