Notify watchdog appropriately when executing a delayed IAccessibleHandler call to correct the focus.
Previously, the watchdog didn't know about these calls, so they weren't subject to possible cancellation. This resulted in unrecovered freezes, for example, when alt+tabbing to an unresponsive process.
core.callLater was introduced to facilitate this. It should be used instead of wx.CallLater for any delayed calls which aren't just for UI.