In Windows Vista, a spurious "Entry Point Not Found" dialog is no longer displayed in several cases such as when starting NVDA from the Desktop shortcut or via the shortcut key.
Py2exe was bundling mpr.dll, which is a Windows dll that differs across different versions of Windows. ShellExecute was loading our copy of this dll instead of the appropriate dll for the system.
This has now been added to the list of system dlls that we explicitly exclude in setup.py.
We also remove mpr.dll when updating, as it will have been installed by an older version.
Fixes #4235.