Load specific app modules for apps hosted by javaw.exe.
If the parent process isn't explorer or cmd, the parent process name will be used, as this is probably a launcher.
Otherwise, the Java class name will be extracted from the command line and a name will be constructed from that. For example, if the class name was com.sun.deploy.panel.ControlPanel, the app name NVDA uses will be javaw_com_sun_deploy_panel_ControlPanel.