When you send a message in Skype, it is no longer read twice.
We were previously using the list item text to find the last message. In recent versions of Skype, this causes problems because the delivery status is included in the text, so the text changes once the message is delivered.
Thankfully, in recent versions of Skype, we can rely on the child ids to detect new messages, which simplifies things a lot and fixes this double reporting problem.
Re #3616.