Skype: Fix complete brokenness of the review message command in some cases.
Sometimes, particularly in Split Window view, we were getting gainFocus and focusEntered events on two different instances of the Conversation object. The second (uninitialised) instance was being used for the review message command.
Fix this by ensuring that we aren't dealing with two different instances. See _isEqual for details.
Re #3210.