Force the mouse text unit resolution for Word documents to sentence if the user has configured the mouse chunk to be line. Re #4852
In 2014.4 expanding to line for anything in MS Word but the caret would expand to sentence. Now expand to line always expands to line, however it can sometimes inappropriately scroll the document. This is very annoying when reading with the mouse.
Therefore, for now force the mouse chunk back to sentence to retain the old behaviour, until the scrolling can be fixed.