When using Baum/HumanWare/APH braille displays with a braille keyboard, braille input no longer stops functioning after pressing another type of key on the display.
When all keys for a group were released, a 0 was still being set in the keysDown dict for that group. However, when testing for braille input, we count the number of groups in keysDown, assuming that groups with no keys down are not present at all.
To fix this, remove a group altogether when all keys are released.
Re #3541.