In Windows 7 and later, text is no longer garbled when typing in certain applications such as Wordpad and Skype with a braille display.
This is related to Rich Edit controls. It occurred because in order to get formatting information in these controls, we have to move the selection (and thus the cursor), which obviously causes problems if typed characters are arriving during this process.
To fix this, we can use ITextDocument. We previously disabled this due to poor performance, but it seems this is fixed in Windows 7 and later. Therefore, use ITextDocument in Windows 7 and later.
Also, the previous code for detecting links in ITextDocument still moved the selection. However, ITextDocument provides for another way to detect links which doesn't do this.
Fixes #4291.