Add support for cross references in word
now able to report cross references as links when reading through text in word.
Browse mode, navigating by links with 'k' does not work
Activating the links do not work.
Probably need more validation to check that the ref is also a link. This
can be done by checking for the presence of \\h in the
`fields.item(1).code.text` result eg: `u' REF _Ref457210120 \\h `