Fix issue where only the last character was selected when using NVDA+f9 and NVDA+f10 in Mozilla editable text controls. (#6306)
When moving just the end of a range, MozillaCompoundTextInfo collapses the end TextInfo. However, the start and end TextInfos are often the same object, so this was causing the start to be lost. If start is end, we now copy end first before tweaking it. Do similarly when just moving the start.