I6312 fix mouse interaction symbol list (#6328)
Fixes #6312 
PR: #6328

The mouse can now be used more effectively in the symbol pronunciation dialog.

Updating the symbol list items as the "change symbol" fields are
changed fixes an issue where the values for the items in the list were
not updated when the mouse was used to select a new item in the list.
When the already selected (and now changed) item was reselected, then
the item in the list WAS updated.
This issue seemed to be caused by the EVT_KILL_FOCUS and the
EVT_LIST_ITEM_FOCUSED being received out of order.

The "change symbol" fields are now disabled when no item is selected (when dialog is
first opened). Seeing the fields disabled helps to indicate that an item
must be selected first. When this happens its clear that the fields
relate to the selected item, as they can be seen to update.