Fix regression in backwards panning introduced with the new focus context presentation settings (issue #7468, PR #7470)
When panning backwards, the end position of the window is set to the previous start position. Thus, the end position of the window belongs to the region that was initially focused hard left. We should not check focusToHardLeft for this region again. regionStart should be less than endPos, and not less than or equal to it.