Remove NVDA service, no longer needed as NVDA will require Windows 7 SP1 or higher (#7625)
* NVDA service: py2exe will no longer build the service module, as the service is only needed on XP and Server 2003.

* Installer: no longer need to start NVDA service, as Windows 7 SP1 or higher is required now.

* NVDA service is hereby removed.

* Just assume Ease of Access is available.

Noted by Leonard de Ruijter: just assume Ease of Access is present.

* NVDA slave: proper copyright header, no longer manages NVDA service.

Added proper copyright header for the slave process. Also, as XP is no longer supported, service launcher is no more.

* Config: simplify secure screen startup by checking if NVDA is installed.

No more support for XP/Vista/7 RTM means:
* No more NVDA service.
* No more check for presence of Ease of Access.
Thus, the secure screen startup checker just returns if NVDA is installed, thus ready to be transformed into a flag in a future release.

* SConstruct: no more service builder.

Reviewed by Mick Curran (NV Access): knew something was incomplete when working on removing NVDA service - SCons should no longer be told to build the service executable, thereby allowing signed builds to proceed.

* NvDA service: include win32api so some add-ons can function properly.

The following add-on is known for using win32api: Golden Cursor. Because of the impact of this add-on, include win32api in 2017.4, to be removed in 2018 as part of Python 3 transition or a major code rewrite.