Touch keyboard support: ability to configure how the key is pressed via hover up (tuch typing) or double-tap (standard typing) (#7316)
* Touch support: add a new map for touch interaction settings such as standard versus touch typing and what not. re #7309.

* Global commands/hover up: press the touch key if touch typing option is enabled.

By default, touch typing will be disabled, but can be enabled by changing touch interaction settings.

* Touchscreen/touch typing: add a new Touch Interaction dialog to configure touch typing mode between touch and standard.

Instead of just calling this 'touchscreen settings', touch interaction settings is a more appropriate title as it may include other interaction methods and settings in the future.

* User guide: document the new touch interaction dialog and added a section on using the touch keyboard.

* Touch interaction: only add this dialog if and only if touchscreen support is in use. re #7309

* Fix translator comment mispelling.

* Touch Interaction: make sure to call the correct super function for the correct dialog class.

* User guide: includethe fact that tablets display touch keyboard when keyboards are undocked. re #7309.

A reminder from Mick Curran (NV Access): for tablets such as Microsoft surface Pro and others, when the computer is in tablet mode or keyboard is undocked, touch keyboard becomes visible and stays on top at all times.

* Touch interaction dialog: add translator comment for the tooltip.

A word from Mick Curran (NV Access) to add translator comments in order for scons checkPot to pass.

* Touch handler and GUI: check if touch is supported, as touch handler might not be ready when initializing GUI subsystem. re #7309.

Touch handler is initialized AFTER GUI, so handler might not be ready at all. To avoid this, a new touchHandler.touchSupported function will be used to check if touch interaction is supported. For now, it is used by GUI, but add-ons and other modules are more than welcome to use it."

* Touch handler: docstring

* Touch handler: use max touches constant (95) as a proper constant. re #7309.

Noted by Mick Curran (NV Access): because the literal '95' is used multiple times, give it a proper label and use it as a constant.