Fix browse mode in Firefox extension popups. (#7809)
Firefox extensions can pop up documents to provide access to their functionality without creating a new browser tab.
For example, both the LastPass and bitwarden password managers pop up a document allowing you to access your vault, add a site, etc. when you press their button on the toolbar.
Previously, browse mode only worked in these documents when you were focused on the document itself.
If you focused something inside the document (e.g. a link, a button or a text box), browse mode functionality ceased to work.
This could happen even if you moved the browse mode cursor, so you would suddenly be "thrown out" of browse mode.

This occurred because the Gecko vbuf code to handle combo box popups was being incorrectly used in this case.
It only checked for Mozilla popup windows, but these popup documents are also popup windows.
It walked to the nearest ancestor with a different HWND, but that meant that the document was skipped.
Thus, anything inside the document was never considered to be part of the document.
Now, the code to deal with combo box popups has been tightened so that it only matches those.