Parse LaTeX in the user guide to MathML (#19304)
Closes #19014
Summary of the issue:

Since #18323, there has been LaTeX included in the User Guide. However, it was not parsed, just left as source.
Description of user facing changes:

The newly added LaTeX is now parsed to MathML, meaning it may be interacted with using MathCAT, and should be displayed in modern browsers.
Description of developer facing changes:

None
Description of development approach:

    Added l2m4m to as a dependency.
    Added its LaTeX2MathMLExtension to the extensions used when converting the user guide.
    Whitelisted the MathML tags that we currently use in the user guide.
    Allowed the display attribute on <math> tags.
    Added an attribute filter callable that strips the display attribute on <math> tags if its value is inline, as this is redundant.