Move mathCATDir to ReadPaths (#19370)
Fixup #18323

### Summary of the issue:

`mathCATDir` is a property of `NVDAState.WritePaths`, even though it is
not writable by non-elevated installed copies, and it doesn't seem like
we even want to write to it.

### Description of user facing changes:

None

### Description of developer facing changes:

`mathCATDir` is now a property of `NVDAState.ReadDirs`.

### Description of development approach:

Moved `mathCATDir` to `NVDAState._ReadDirs`.
Searched for usages of `mathCATDir` in `source/` and updated them.

### Testing strategy:

Ran from source and interacted with MathML in Firefox.

### Known issues with pull request:

None